requirements.txt
customtkinter
CTkMessagebox
Pillow>=10.0.0
